Refactor panfs caching.
We now use volume quotas. This commit uses the panfs XML API to fetch
all volume quotas to cache. We also set the file indode limits using
user/group quotas. As a simple hack we just set the used inodes for a
volume to be that of the largest group.